Creating Scripting contrib module to centralize the less secure code related
to scripts.
