+++ gradle/maven/defaults-maven.gradle
@@ -66,6 +66,7 @@ configure(rootProject) {
         ":solr:contrib:langid",
         ":solr:contrib:jaegertracer-configurator",
         ":solr:contrib:prometheus-exporter",
+        ":solr:contrib:scripting",
         ":solr:test-framework",
     ]
