+++ solr/contrib/scripting/src/test-files/solr/collection1/conf/addfields.updateprocessor.js
