+++ solr/contrib/scripting/src/test-files/solr/collection1/conf/currency.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" ?>
+<!--
+ Licensed to the Apache Software Foundation (ASF) under one or more
+ contributor license agreements.  See the NOTICE file distributed with
+ this work for additional information regarding copyright ownership.
+ The ASF licenses this file to You under the Apache License, Version 2.0
+ (the "License"); you may not use this file except in compliance with
+ the License.  You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+-->
+
+<!-- Example exchange rates file,
+
+
+     NOTE: this data must be kept in sync with ./open-exchange-rates.json
+-->
+
+<currencyConfig version="1.0">
+  <rates>
+    <!-- Fake rates for testing -->
+    <rate from="USD" to="JPY" rate="81.29"/>
+    <rate from="USD" to="EUR" rate="2.5"/>
+    <rate from="USD" to="GBP" rate="0.5"/>
+    <rate from="EUR" to="GBP" rate="0.5"/>
+    <rate from="MXN" to="USD" rate="0.5"/>
+
+    <!-- Asymmetric rate -->
+    <rate from="EUR" to="USD" rate="0.5"/>
+  </rates>
+</currencyConfig>
