+++ solr/contrib/scripting/src/test-files/solr/collection1/conf/trivial.updateprocessor0.js
