+++ solr/contrib/scripting/src/test-files/solr/collection1/conf/trivial.updateprocessor1.js
