+++ solr/contrib/scripting/src/test/org/apache/solr/scripting/update/ScriptEngineTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.update.processor;
+package org.apache.solr.scripting.update;
 
 import org.apache.lucene.util.Constants;
