+++ solr/solr-ref-guide/src/configuring-solrconfig-xml.adoc
@@ -8,6 +8,7 @@
     query-settings-in-solrconfig, \
     requestdispatcher-in-solrconfig, \
     update-request-processors, \
+    scripting-update-processor, \    
     codec-factory
 
 // Licensed to the Apache Software Foundation (ASF) under one
