+++ solr/solr-ref-guide/src/major-changes-in-solr-9.adoc
@@ -144,6 +144,8 @@ _(raw; not yet edited)_
 
 * SOLR-14972: The default port of prometheus exporter has changed from 9983 to 8989, so you may need to adjust your configuration after upgrade.
 
+* SOLR-14067: StatelessScriptingUpdateProcessor moved to /contrib/scripting/ module and renamed to ScriptingUpdateProcessor.
+
 === Upgrade Prerequisites in Solr 9
 
 * Upgrade all collections in stateFormat=1 to stateFormat=2 *before* upgrading to Solr 9, as Solr 9 does not support the
